+++ apps/block_scout_web/assets/js/lib/history_chart.js
@@ -1,4 +1,5 @@
 import $ from 'jquery'
+import Cookies from 'js-cookie'
 import { Chart, LineController, LineElement, PointElement, LinearScale, TimeScale, Title, Tooltip } from 'chart.js'
 import 'chartjs-adapter-luxon'
 import humps from 'humps'
@@ -17,7 +18,7 @@ const grid = {
 }
 
 function getTxChartColor () {
-  if (localStorage.getItem('current-color-mode') === 'dark') {
+  if (Cookies.get('chakra-ui-color-mode') === 'dark') {
     return sassVariables.dashboardLineColorTransactionsDarkTheme
   } else {
     return sassVariables.dashboardLineColorTransactions
@@ -25,7 +26,7 @@ function getTxChartColor () {
 }
 
 function getPriceChartColor () {
-  if (localStorage.getItem('current-color-mode') === 'dark') {
+  if (Cookies.get('chakra-ui-color-mode') === 'dark') {
     return sassVariables.dashboardLineColorPriceDarkTheme
   } else {
     return sassVariables.dashboardLineColorPrice
@@ -33,7 +34,7 @@ function getPriceChartColor () {
 }
 
 function getMarketCapChartColor () {
-  if (localStorage.getItem('current-color-mode') === 'dark') {
+  if (Cookies.get('chakra-ui-color-mode') === 'dark') {
     return sassVariables.dashboardLineColorMarketDarkTheme
   } else {
     return sassVariables.dashboardLineColorMarket